refactor(android): split core code from rn code (#1185) (#1185)
# Pull Request Checklist

- [ ] Did you add new tests and confirm existing tests pass? (`yarn test`)
- [ ] Did you update relevant docs? (docs are found in the `site` folder, and guidelines for updating/adding docs can be found in the [contribution guide](https://github.com/alchemyplatform/aa-sdk/blob/main/CONTRIBUTING.md))
- [ ] Do your commits follow the [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/) standard?
- [ ] Does your PR title also follow the [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/) standard?
- [ ] If you have a breaking change, is it [correctly reflected in your commit message](https://www.conventionalcommits.org/en/v1.0.0/#examples)? (e.g. `feat!: breaking change`)
- [ ] Did you run lint (`yarn lint:check`) and fix any issues? (`yarn lint:write`)
- [ ] Did you follow the [contribution guidelines](https://github.com/alchemyplatform/aa-sdk/blob/main/CONTRIBUTING.md)?

<!-- start pr-codex -->

---

## PR-Codex overview
This PR introduces new error handling classes and refactors the `NativeTEKStamperModule` to utilize a new `TEKStamper` class for better encapsulation of functionality related to TEK management, improving code organization and clarity.

### Detailed summary
- Added `NoTEKException`, `StamperNotInitializedException`, and `NoInjectedBundleException` classes.
- Refactored `KeyExtensions.kt` and `TEKManager.kt` to use the new error classes.
- Introduced `TEKStamper` class to handle TEK operations.
- Updated `NativeTEKStamperModule` to delegate TEK operations to `TEKStamper`.
- Removed redundant code from `NativeTEKStamperModule`.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your question}`

<!-- end pr-codex -->
