+++ account-kit/rn-signer/android/src/main/java/com/accountkit/reactnativesigner/core/errors/NoInjectedBundleException.kt
@@ -1,0 +1,3 @@
+package com.accountkit.reactnativesigner.core.errors
+
+class NoInjectedBundleException: IllegalStateException("No injected bundle, did you complete auth?")
