+++ account-kit/rn-signer/android/src/main/java/com/accountkit/reactnativesigner/core/errors/NoTEKException.kt
@@ -1,0 +1,3 @@
+package com.accountkit.reactnativesigner.core.errors
+
+class NoTEKException: IllegalStateException("No TEK found!")
