+++ account-kit/rn-signer/android/src/main/java/com/accountkit/reactnativesigner/core/errors/StamperNotInitialized.kt
@@ -1,0 +1,3 @@
+package com.accountkit.reactnativesigner.core.errors
+
+class StamperNotInitializedException: IllegalStateException("Stamper has not been initialized")
